Books Model (#3)
The books model is added to the application.
The books model covers the input model, information model and response model, allowing users input search queries for books and get a response from the query of the database.
Tests for the books model are also added, enforcing standard codebase development practices.